添加伤害皮肤 部分分辨率显示修复

+++ ezorsia/MapleClientCollectionTypes/CInPacket.h
@@ -20,7 +20,7 @@ public:
 public:
 	char Decode1();
 	short Decode2();
-	unsigned long Decode4();
+	int Decode4();
 	int DecodeBuffer(PVOID out, int len);
 	unsigned char* DecodeStr(ZXString<char>* p);
 	int DecryptData(int unknow);
